array/array_lc76_test.go
package array

import (
	"testing"
)

func Test_minWindow(t *testing.T) {
	type args struct {
		s string
		t string
	}
	tests := []struct {
		name string
		args args
		want string
	}{
		{
			name: "lc case1",
			args: args{
				s: "ADOBECODEBANC",
				t: "ABC",
			},
			want: "BANC",
		},
		{
			name: "lc case2",
			args: args{
				s: "a",
				t: "a",
			},
			want: "a",
		},

		{
			name: "lc case3",
			args: args{
				s: "a",
				t: "aa",
			},
			want: "",
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := minWindow(tt.args.s, tt.args.t); got != tt.want {
				t.Errorf("minWindow() = %v, want %v", got, tt.want)
			}
		})
	}
}

func Test_minWindowPV1(t *testing.T) {
	type args struct {
		s string
		t string
	}
	tests := []struct {
		name string
		args args
		want string
	}{
		{
			name: "lc case1",
			args: args{
				s: "ADOBECODEBANC",
				t: "ABC",
			},
			want: "BANC",
		},
		{
			name: "lc case2",
			args: args{
				s: "a",
				t: "a",
			},
			want: "a",
		},

		{
			name: "lc case3",
			args: args{
				s: "a",
				t: "aa",
			},
			want: "",
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := minWindowPV1(tt.args.s, tt.args.t); got != tt.want {
				t.Errorf("minWindowPV1() = %v, want %v", got, tt.want)
			}
		})
	}
}
